Ball is moving and rolling!

+++ src/GolfEngine/GameManagement/Levels/Level.hpp
@@ -79,7 +79,7 @@ namespace GolfEngine {
              *
              * @param dt Time since last frame update (in ms)
              */
-            void frameUpdate(float dt);
+            void frameUpdate(uint dt);
         private:
             GolfEngine::Vector2 target;
